Maked chart vertically split
