+++ src/components/Controllers/Chart/Editor.tsx
@@ -1,18 +1,13 @@
 import * as React from 'react'
-import { VegaLite } from 'react-vega'
-import Box from '@mui/material/Box'
-import Preset from './Preset'
-import { useStore } from './store'
+import Columns from '../../Parts/Columns'
+import Target from './Target'
+import Source from './Source'
 
 export default function Editor() {
-  const chart = useStore((state) => state.chart)
-  const fields = useStore((state) => state.fields)
-  // const updateState = useStore((state) => state.updateState)
-  if (!fields) return null
   return (
-    <Box sx={{ mt: 2 }}>
-      {!chart && <Preset />}
-      {!!chart && <VegaLite spec={chart} width={1000} height={550} />}
-    </Box>
+    <Columns height="100%">
+      <Source />
+      <Target />
+    </Columns>
   )
 }
