+++ src/components/Controllers/Chart/Preset.tsx
@@ -1,13 +1,10 @@
 import * as React from 'react'
 import Columns from '../../Parts/Columns'
-import VerticalTabs from '../../Parts/Tabs/Vertical'
 
 export default function Preset() {
   return (
     <Columns spacing={3} layout={[9, 3]}>
-      <VerticalTabs labels={['Bar']}>
-        <PresetBar />
-      </VerticalTabs>
+      <PresetBar />
     </Columns>
   )
 }
